Double-Letter Case, Score, Summary, Cleanup
Double-Letter Case: Fixed the case where a word ends in double letters,
and the second to last letter is the one to be filled in.

Score: Added a score() function in php--all it does now is make every
word worth 9001.  Handled the php/js handoff of the score, and made the
score show up as long as the word shows up on the board.

Summary: Added php to summarize the options before they are
listed--still need to be fleshed out.

Cleanup: Removed a lot of debug code, but just commented out debug for
the more complicated functions.